+++ cli/cmd/sync.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"strings"
+	"time"
 
 	"github.com/cloudquery/plugin-sdk/clients"
 	"github.com/cloudquery/plugin-sdk/specs"
@@ -36,7 +37,7 @@ func NewCmdSync() *cobra.Command {
 
 func sync(cmd *cobra.Command, args []string) error {
 	ctx := cmd.Context()
-
+	log.Info().Strs("args", args).Msg("Loading spec(s)")
 	fmt.Printf("Loading spec(s) from %s\n", strings.Join(args, ", "))
 	specReader, err := specs.NewSpecReader(args)
 	if err != nil {
@@ -64,13 +65,25 @@ func sync(cmd *cobra.Command, args []string) error {
 }
 
 func syncConnection(ctx context.Context, sourceSpec specs.Source, destinationsSpecs []specs.Destination) error {
-	sourceClient, err := clients.NewSourceClient(ctx, sourceSpec.Registry, sourceSpec.Path, sourceSpec.Version)
+	destinationNames := make([]string, len(destinationsSpecs))
+	for i := range destinationsSpecs {
+		destinationNames[i] = destinationsSpecs[i].Name
+	}
+	syncTime := time.Now().UTC()
+
+	log.Info().Str("source", sourceSpec.Name).Strs("destinations", destinationNames).Time("sync_time", syncTime).Msg("Start sync")
+	defer log.Info().Str("source", sourceSpec.Name).Strs("destinations", destinationNames).Time("sync_time", syncTime).Msg("End sync")
+
+	sourceClient, err := clients.NewSourceClient(ctx, sourceSpec.Registry, sourceSpec.Path, sourceSpec.Version,
+		clients.WithSourceLogger(log.Logger),
+	)
 	if err != nil {
 		return fmt.Errorf("failed to get source plugin client for %s: %w", sourceSpec.Name, err)
 	}
 	defer func() {
-		if err := sourceClient.Close(); err != nil {
-			fmt.Println("failed to close source client: ", err)
+		if err := sourceClient.Terminate(); err != nil {
+			log.Error().Err(err).Msg("Failed to terminate source client")
+			fmt.Println("failed to terminate source client: ", err)
 		}
 	}()
 
@@ -82,14 +95,17 @@ func syncConnection(ctx context.Context, sourceSpec specs.Source, destinationsSp
 	defer func() {
 		for _, destClient := range destClients {
 			if destClient != nil {
-				if err := destClient.Close(); err != nil {
-					fmt.Println("failed to close destination client: ", err)
+				if err := destClient.Terminate(); err != nil {
+					log.Error().Err(err).Msg("Failed to terminate destination client")
+					fmt.Println("failed to terminate destination client: ", err)
 				}
 			}
 		}
 	}()
 	for i, destinationSpec := range destinationsSpecs {
-		destClients[i], err = clients.NewDestinationClient(ctx, destinationSpec.Registry, destinationSpec.Path, destinationSpec.Version)
+		destClients[i], err = clients.NewDestinationClient(ctx, destinationSpec.Registry, destinationSpec.Path, destinationSpec.Version,
+			clients.WithDestinationLogger(log.Logger),
+		)
 		if err != nil {
 			return fmt.Errorf("failed to create destination plugin client for %s: %w", destinationSpec.Name, err)
 		}
@@ -107,11 +123,12 @@ func syncConnection(ctx context.Context, sourceSpec specs.Source, destinationsSp
 	}
 
 	resources := make(chan []byte)
-	g, ctx := errgroup.WithContext(ctx)
+	g, gctx := errgroup.WithContext(ctx)
+	log.Info().Str("source", sourceSpec.Name).Strs("destinations", sourceSpec.Destinations).Msg("Start fetching resources")
 	fmt.Println("Starting sync for: ", sourceSpec.Name, "->", sourceSpec.Destinations)
 	g.Go(func() error {
 		defer close(resources)
-		if err := sourceClient.Sync(ctx, sourceSpec, resources); err != nil {
+		if err := sourceClient.Sync(gctx, sourceSpec, resources); err != nil {
 			return fmt.Errorf("failed to sync source %s: %w", sourceSpec.Name, err)
 		}
 		return nil
@@ -123,6 +140,7 @@ func syncConnection(ctx context.Context, sourceSpec specs.Source, destinationsSp
 		progressbar.OptionShowIts(),
 		progressbar.OptionSetElapsedTime(true),
 		progressbar.OptionShowCount(),
+		progressbar.OptionClearOnFinish(),
 	)
 	failedWrites := uint64(0)
 	totalResources := 0
@@ -132,8 +150,8 @@ func syncConnection(ctx context.Context, sourceSpec specs.Source, destinationsSp
 		g.Go(func() error {
 			var destFailedWrites uint64
 			var err error
-			if destFailedWrites, err = destClients[i].Write(ctx, destSubscriptions[i]); err != nil {
-				log.Error().Err(err).Msgf("failed to write for %s->%s", sourceSpec.Name, destination)
+			if destFailedWrites, err = destClients[i].Write(gctx, sourceSpec.Name, syncTime, destSubscriptions[i]); err != nil {
+				return fmt.Errorf("failed to write for %s->%s: %w", sourceSpec.Name, destination, err)
 			}
 			failedWrites += destFailedWrites
 			return nil
@@ -146,8 +164,8 @@ func syncConnection(ctx context.Context, sourceSpec specs.Source, destinationsSp
 			_ = bar.Add(1)
 			for i := range destSubscriptions {
 				select {
-				case <-ctx.Done():
-					return ctx.Err()
+				case <-gctx.Done():
+					return gctx.Err()
 				case destSubscriptions[i] <- resource:
 				}
 			}
@@ -160,10 +178,18 @@ func syncConnection(ctx context.Context, sourceSpec specs.Source, destinationsSp
 
 	if err := g.Wait(); err != nil {
 		_ = bar.Finish()
-		return fmt.Errorf("failed to fetch resources: %w", err)
+		return err
+	}
+	summary, err := sourceClient.GetSyncSummary(ctx)
+	if err != nil {
+		return fmt.Errorf("failed to get sync summary: %w", err)
 	}
 	_ = bar.Finish()
+	tt := time.Since(syncTime)
+
 	fmt.Println("Sync completed successfully.")
-	fmt.Printf("Summary: Resources: %d\n", totalResources)
+	fmt.Printf("Summary: resources: %d, errors: %d, panic: %d, failed_writes: %d, time: %s\n", totalResources, summary.Errors, summary.Panics, failedWrites, tt.Truncate(time.Second).String())
+	log.Info().Str("source", sourceSpec.Name).Strs("destinations", sourceSpec.Destinations).
+		Int("resources", totalResources).Uint64("errors", summary.Errors).Uint64("panic", summary.Panics).Uint64("failed_writes", failedWrites).Float64("time_took", tt.Seconds()).Msg("Sync completed successfully")
 	return nil
 }