+++ plugins/source/aws/codegen/recipes/route53.go
@@ -1,8 +1,12 @@
 package recipes
 
 import (
+	"reflect"
+	"strings"
+
 	"github.com/aws/aws-sdk-go-v2/service/route53/types"
-	"github.com/aws/aws-sdk-go/service/route53domains"
+	"github.com/aws/aws-sdk-go-v2/service/route53domains"
+	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/route53/models"
 	"github.com/cloudquery/plugin-sdk/codegen"
 	"github.com/cloudquery/plugin-sdk/schema"
 )
@@ -28,9 +32,10 @@ func Route53Resources() []*Resource {
 		},
 
 		{
-			SubService: "domains",
-			Struct:     &route53domains.GetDomainDetailOutput{},
-			SkipFields: []string{"DomainName", "_"},
+			SubService:          "domains",
+			Struct:              &route53domains.GetDomainDetailOutput{},
+			SkipFields:          []string{"DomainName", "_"},
+			PreResourceResolver: "getDomain",
 			ExtraColumns: []codegen.ColumnDefinition{
 				{
 					Name:     "account_id",
@@ -80,7 +85,7 @@ func Route53Resources() []*Resource {
 
 		{
 			SubService: "hosted_zones",
-			Struct:     &types.HostedZone{},
+			Struct:     &models.Route53HostedZoneWrapper{},
 			SkipFields: []string{"ARN"},
 			ExtraColumns: append(
 				defaultAccountColumns,
@@ -174,7 +179,7 @@ func Route53Resources() []*Resource {
 		{
 			SubService: "traffic_policy_versions",
 			Struct:     &types.TrafficPolicy{},
-			SkipFields: []string{"Version", "Id"},
+			SkipFields: []string{"Version", "Id", "Document"},
 			ExtraColumns: append(
 				defaultAccountColumns,
 				[]codegen.ColumnDefinition{
@@ -196,6 +201,11 @@ func Route53Resources() []*Resource {
 						Resolver: `schema.PathResolver("Version")`,
 						Options:  schema.ColumnCreationOptions{PrimaryKey: true},
 					},
+					{
+						Name:     "document",
+						Type:     schema.TypeJSON,
+						Resolver: `client.MarshaledJsonResolver("Document")`,
+					},
 				}...),
 		},
 	}
@@ -204,6 +214,10 @@ func Route53Resources() []*Resource {
 	for _, r := range resources {
 		r.Service = "route53"
 		r.Multiplex = "client.AccountMultiplex"
+		structName := reflect.ValueOf(r.Struct).Elem().Type().Name()
+		if strings.Contains(structName, "Wrapper") {
+			r.UnwrapEmbeddedStructs = true
+		}
 	}
 	return resources
 }