+++ plugins/source/aws/resources/services/cognito/user_pool_identity_providers.go
@@ -9,9 +9,10 @@ import (
 
 func UserPoolIdentityProviders() *schema.Table {
 	return &schema.Table{
-		Name:      "aws_cognito_user_pool_identity_providers",
-		Resolver:  fetchCognitoUserPoolIdentityProviders,
-		Multiplex: client.ServiceAccountRegionMultiplexer("cognito-identity"),
+		Name:                "aws_cognito_user_pool_identity_providers",
+		Resolver:            fetchCognitoUserPoolIdentityProviders,
+		PreResourceResolver: getUserPoolIdentityProvider,
+		Multiplex:           client.ServiceAccountRegionMultiplexer("cognito-identity"),
 		Columns: []schema.Column{
 			{
 				Name:     "account_id",