+++ plugins/source/aws/resources/services/fsx/data_repository_associations.go
@@ -57,9 +57,24 @@ func DataRepositoryAssociations() *schema.Table {
 				Resolver: schema.PathResolver("DataRepositoryPath"),
 			},
 			{
+				Name:     "data_repository_subdirectories",
+				Type:     schema.TypeStringArray,
+				Resolver: schema.PathResolver("DataRepositorySubdirectories"),
+			},
+			{
 				Name:     "failure_details",
 				Type:     schema.TypeJSON,
 				Resolver: schema.PathResolver("FailureDetails"),
+			},
+			{
+				Name:     "file_cache_id",
+				Type:     schema.TypeString,
+				Resolver: schema.PathResolver("FileCacheId"),
+			},
+			{
+				Name:     "file_cache_path",
+				Type:     schema.TypeString,
+				Resolver: schema.PathResolver("FileCachePath"),
 			},
 			{
 				Name:     "file_system_id",
@@ -80,6 +95,11 @@ func DataRepositoryAssociations() *schema.Table {
 				Name:     "lifecycle",
 				Type:     schema.TypeString,
 				Resolver: schema.PathResolver("Lifecycle"),
+			},
+			{
+				Name:     "nfs",
+				Type:     schema.TypeJSON,
+				Resolver: schema.PathResolver("NFS"),
 			},
 			{
 				Name:     "s3",